feat: implement entity field extension
