+++ engines/auth_by/app/graphql/auth_by/mutations/sign_in.rb
@@ -16,18 +16,18 @@ module AuthBy
 
       argument :input, SignInInput, "Credentials input", required: true
 
-      field :user, CoreBy::Types::User, null: true
+      field :user, CoreBy::Types::User, null: true, extensions: [CoreBy::Schema::EntityFieldExt]
       field :access_token, String, "JWT access token", null: true
       field :refresh_token, String, "JWT refresh token", null: true
 
       def resolve(input:)
-        core_user = CoreBy::SDK::Users.find_by_email(input.email)
-        user = User.find_by_id(core_user.id) if core_user && !core_user.disabled?
+        entity_user = CoreBy::SDK::Users.find_by_email(input.email)
+        user = User.find_by_id(entity_user.id) if entity_user && !entity_user.disabled?
         fail_with!(:invalid, "Unauthenticated") unless user&.valid_password?(input.password)
 
         tokens = user.generate_jwt_tokens
         {
-          user: core_user.to_record,
+          user: entity_user,
           access_token: tokens.fetch(:access),
           refresh_token: tokens.fetch(:refresh)
         }
