+++ engines/auth_by/app/graphql/auth_by/mutations/sign_up.rb
@@ -19,7 +19,7 @@ module AuthBy
 
       argument :input, SignUpInput, required: true
 
-      field :user, CoreBy::Types::User, null: true
+      field :user, CoreBy::Types::User, null: true, extensions: [CoreBy::Schema::EntityFieldExt]
       field :access_token, String, "JWT access token", null: true
       field :refresh_token, String, "JWT refresh token", null: true
       field :errors, CoreBy::Types::ValidationErrors, "Errors when a user cannot be created", null: true
@@ -29,10 +29,10 @@ module AuthBy
 
         if form.save
           tokens = form.user.generate_jwt_tokens
-          core_user = CoreBy::SDK::Users.find_by_id(form.user.id)
+          entity_user = CoreBy::SDK::Users.find_by_id(form.user.id)
 
           {
-            user: core_user.to_record,
+            user: entity_user,
             access_token: tokens[:access],
             refresh_token: tokens[:refresh]
           }
